requirements.txt
BeautifulSoup==3.2.0
Brlapi==0.5.6
GnuPGInterface==0.3.2
Jinja2==2.6
Magic-file-extensions==0.2
Mako==0.5.0
MarkupSafe==0.15
MySQL-python==1.2.3
PAM==0.4.2
PIL==1.1.7
PsychoPy==1.77.01
PyAudio==0.2.4
PyOpenGL==3.0.1b2
Pygments==1.4
Pyste==0.9.10
SOAPpy==0.12.0
Sphinx==1.1.3
Twisted-Core==11.1.0
Twisted-Names==11.1.0
Twisted-Web==11.1.0
UniConvertor==1.1.4
adium-theme-ubuntu==0.3.2
ansicolors==1.0.2
apt-xapian-index==0.44
apturl==0.5.1ubuntu3
argparse==1.2.1
bzr==2.5.1
bzr-builddeb==2.7.0
ccsm==0.9.5.92
chainmap==1.0.2
chardet==2.0.1
codegen==1.0
combomethod==1.0.6
command-not-found==0.2.44
compizconfig-python==0.9.5.94
configglue==1.0
configobj==4.7.2
debtagshw==0.1
decorator==3.3.2
defer==1.0.2
devscripts==2.11.6ubuntu1.6
dirspec==3.0.0
distribute==0.6.24
distro-info==0.8.2
docutils==0.8.1
dulwich==0.8.5
duplicity==0.6.18
fastimport==0.9.0
feedparser==5.1
fpconst==0.7.2
httplib2==0.7.2
ipython==0.12.1
jockey==0.9.7
keyring==0.9.2
language-selector==0.1
launchpadlib==1.9.12
lazr.restfulclient==0.12.0
lazr.uri==1.0.3
lockfile==0.8
logilab-astng==0.23.0
logilab-common==0.57.1
louis==2.3.0
lxml==2.3.2
matplotlib==1.1.1rc
mementos==1.2.2
mercurial==2.0.2
nose==1.1.2
nulltype==2.2.5
numexpr==1.4.2
numpy==1.8.1
nvidia-common==0.0.0
oauth==1.0.1
onboard==0.97.1
oneconf==0.2.8.1
openpyxl==1.7.0
openshot==1.4.0
options==1.4.3
paramiko==1.7.7.1
pep8==0.6.1
pexpect==2.3
piston-mini-client==0.7.2
protobuf==2.4.1
psutil==0.4.1
pyOpenSSL==0.12
pyalsaaudio==0.8.2
pycrypto==2.4.1
pycuda==2012.1
pycups==1.9.61
pycurl==7.19.0
pydns==2.3.6
pyflakes==0.5.0
pygame==1.9.1release
pyglet==1.1.4
pygpgme==0.2
pyinotify==0.9.2
pylint==0.25.0
pyo==0.6.6
pyparsing==1.5.2
pyserial==2.5
pysmbc==1.0.13
python-apt==0.8.3ubuntu7.3
python-dateutil==1.5
python-debian==0.1.21ubuntu1
python-debianbts==1.10
python-virtkey==0.60.0
pytools==2011.5
pytz==2012c
pyxdg==0.19
pyzmq==2.1.11
rabbitvcs==0.16.0
reportbug==6.3.1ubuntu1.1
reportlab==2.5
rhythmbox-ubuntuone==4.2.0
rope==0.9.2
say==1.4.2
scipy==0.9.0
screen-resolution-extra==0.0.0
sessioninstaller==0.0.0
show==1.2.6
simplegeneric==0.7
simplejson==2.3.2
simplere==1.2.8
six==1.10.0
software-center-aptd-plugins==0.0.0
spyder==2.2.5
system-service==0.1.6
tables==2.3.1
tornado==2.1
uTidylib==0.2
ubuntu-dev-tools==0.141ubuntu0.1
ubuntuone-couch==0.3.0
ubuntuone-installer==3.0.2
ubuntuone-storage-protocol==3.0.2
ufw==0.31.1-1
unattended-upgrades==0.1
unity-lens-video==0.3.5
unity-scope-video-remote==0.3.5
usb-creator==0.2.23
virtualenv==1.7.1.2
wadllib==1.3.0
wsgiref==0.1.2
wxPython==2.8.12.1
wxPython-common==2.8.12.1
xdiagnose==2.5.3
xkit==0.0.0
zope.interface==3.6.1
